http://support.bimserver.org/bimserver/topics/how_to_check_the_result_of_compileandrun_when_isrunoke_is_not_available

+++ BimWebServer/src/org/bimserver/web/JspHelper.java
@@ -244,7 +244,7 @@ public class JspHelper {
 	}
 
 	public static SClashDetectionSettings createSClashDetectionSettings(HttpServletRequest request) {
-		float margin = Float.parseFloat(request.getParameter("margin"));
+		double margin = Double.parseDouble(request.getParameter("margin"));
 		SClashDetectionSettings sClashDetectionSettings = new SClashDetectionSettings();
 		sClashDetectionSettings.setMargin(margin);
 		String[] ignoredSplit = request.getParameter("ignored").split(";");
